requirements.txt
Flask
flask_cors
pytest
pytest-flask
peewee
redis
psycopg2-binary